+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   attribute in addition to a node name ([rhbz#1261116])
 - `pcs cluster cib-push` allows to push a diff obtained internally by comparing
   CIBs in specified files ([rhbz#1404233])
+- Added flags `--wait`, `--disabled`, `--group`, `--after`, `--before` into
+  the command `pcs stonith create`
 
 ### Changed
 - `pcs node [un]standby` and `pcs node [un]maintenance` is now atomic even if